Fixes: Agrego condición de que el directorio exista antes de tratar de borrarlo con shutil

+++ web/seal/view/teacher/delivery.py
@@ -80,7 +80,8 @@ def browse(request, iddelivery, file_to_browse=None):
 @login_required
 def explore(request, iddelivery):
     if(len(request.user.teacher_set.all()) > 0): # if an authenticated user "accidentally" access this section, he doesn't get an exception
-        shutil.rmtree(os.path.join(managepath.get_instance().get_temporary_files_path(), str(iddelivery)))
+        if(os.path.exists(os.path.join(managepath.get_instance().get_temporary_files_path(), str(iddelivery)))):
+            shutil.rmtree(os.path.join(managepath.get_instance().get_temporary_files_path(), str(iddelivery)))
         return browse(request, iddelivery)
     else:
         return HTTP_401_UNAUTHORIZED_RESPONSE
